* Fixes para modificar el proveedor del precio de boletin de supplier a Affiliate * Correccion en el schema de precios de proveedores * Correcciones varias al delete de affiliates a la creación de affiliates, etc. originados por la unificación de lso proveedores * Simplificación de estilos relativos a anchod e pantalla

+++ vialidad/WEB-INF/classes/modules/affiliates/actions/AffiliatesDoDeleteAction.php
@@ -18,14 +18,16 @@ class AffiliatesDoDeleteAction extends BaseAction {
 
 		$module = "Affiliates";
 
-		$affiliate = AffiliatePeer::get($_POST["id"]);
+		$affiliate = AffiliateQuery::create()->findOneById($_GET["id"]);
 
-		if ($affiliate->delete())
-			return $mapping->findForwardConfig('success');
+		if (!empty($affiliate)) {
+			if ($affiliate->delete())
+				return $mapping->findForwardConfig('success');
+			else
+				return $mapping->findForwardConfig('failure');
+			}
 		else
-			return $mapping->findForwardConfig('failure');
-
-		return $mapping->findForwardConfig('success');
+			return $mapping->findForwardConfig('success');
 	}
 
 }
